refactor(exports): move ITop10 and ITimeEntry
