+++ src/lev/index.ts
@@ -1,4 +1,4 @@
 export { default as ElmaObject, Gravity, ObjectType } from './ElmaObject';
 export { default as Picture } from './Picture';
 export { default as Polygon } from './Polygon';
-export { default as Level, ITimeEntry, ITop10, Version } from './Level';
+export { default as Level, Version } from './Level';
